Fix bugs

+++ librstm/src/startup.rs
@@ -13,24 +13,24 @@ pub unsafe extern fn reset_handler() {
 
 unsafe fn init_data() {
     extern {
-        static __data_load: u32;
-        static mut __data_start: u32;
-        static __data_end: u32;
+        static __data_load: u8;
+        static mut __data_start: u8;
+        static __data_end: u8;
     }
-    let src = &__data_load as *const u32;
-    let dst = &mut __data_start as *mut u32;
-    let end = &__data_end as *const u32;
+    let src = &__data_load as *const u8;
+    let dst = &mut __data_start as *mut u8;
+    let end = &__data_end as *const u8;
     let count = (end as usize) - (dst as usize);
     ptr::copy_nonoverlapping(src, dst, count);
 }
 
 unsafe fn init_bss() {
     extern {
-        static mut __bss_start: u32;
-        static __bss_end: u32;
+        static mut __bss_start: u8;
+        static __bss_end: u8;
     }
-    let dst = &mut __bss_start as *mut u32;
-    let end = &__bss_end as *const u32;
+    let dst = &mut __bss_start as *mut u8;
+    let end = &__bss_end as *const u8;
     let count = (end as usize) - (dst as usize);
     ptr::write_bytes(dst, 0, count);
 }
